Added an enter button on splashscreen. Moved SplashScreen fragment to fragment package.
