+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,9 @@
         android:theme="@style/Theme.SaveBart">
         <activity
             android:name=".MainActivity"
-            android:exported="true">
+            android:exported="true"
+            android:configChanges="orientation"
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
